+++ src/views/main/components/make-quiz-drawer.tsx
@@ -10,10 +10,13 @@ import { useSession } from 'next-auth/react'
 import { useEffect, useMemo, useState } from 'react'
 import Image from 'next/image'
 import icons from '@/constants/icons'
-import { FolderDrawer, DocumentDrawer } from './make-quiz-drawer-content'
-import { QuizCountSelector } from './make-quiz-dialog-selector'
+import { FolderDrawer, DocumentDrawer } from './make-quiz-drawer-contents'
+import { QuizCountSelector } from './make-quiz-dialog-selectors'
+import Div100vh from 'react-div-100vh'
+import Loading from '@/shared/components/loading'
 
 interface Props {
+  startedCreate: boolean
   categories: CategoryDTO[]
   handleCreateQuizzes: ({ documentIds, count }: { documentIds: number[]; count: number }) => void
   closeDrawer: () => void
@@ -23,6 +26,7 @@ interface Props {
 
 // MakeQuizDrawer 컴포넌트
 const MakeQuizDrawer = ({
+  startedCreate,
   categories,
   handleCreateQuizzes,
   closeDrawer,
@@ -81,91 +85,102 @@ const MakeQuizDrawer = ({
   }, [documentLength])
 
   return (
-    <div className="px-[20px]">
-      <div className="relative h-[48px]">
-        <Button variant="ghost" size="icon" className="ml-[-12px]" onClick={() => closeDrawer()}>
-          <X />
-        </Button>
-        <div className="center text-body1-bold text-gray-09">
-          {quizType === 'MIX_UP' ? 'O/X 퀴즈' : '객관식 퀴즈'}
-        </div>
-      </div>
-
-      <h3 className="text-h3-bold text-gray-09">
-        원하는 폴더와 노트,
-        <br />
-        퀴즈 수를 선택해주세요
-      </h3>
-
-      <div className="mt-[48px]">
-        <div className="flex items-center gap-[27px]">
-          <div className="flex flex-col gap-[10px]">
-            <div className="w-[52px] shrink-0 text-body2-medium text-gray-08">폴더</div>
-            <FolderDrawer
-              categories={categories}
-              step={step}
-              setStep={setStep}
-              folder={folder}
-              doc={doc}
-            />
+    <Div100vh>
+      {startedCreate ? (
+        <Loading center />
+      ) : (
+        <div className="px-[20px]">
+          <div className="relative h-[48px]">
+            <Button
+              variant="ghost"
+              size="icon"
+              className="ml-[-12px]"
+              onClick={() => closeDrawer()}
+            >
+              <X />
+            </Button>
+            <div className="center text-body1-bold text-gray-09">
+              {quizType === 'MIX_UP' ? 'O/X 퀴즈' : '객관식 퀴즈'}
+            </div>
           </div>
 
-          <div className="flex flex-col gap-[10px]">
-            <div className="w-[52px] shrink-0 text-body2-medium text-gray-08">노트</div>
-            <DocumentDrawer
-              categories={categories}
-              step={step}
-              setStep={setStep}
-              folder={folder}
-              doc={doc}
-            />
+          <h3 className="text-h3-bold text-gray-09">
+            원하는 폴더와 노트,
+            <br />
+            퀴즈 수를 선택해주세요
+          </h3>
+
+          <div className="mt-[48px]">
+            <div className="flex items-center gap-[27px]">
+              <div className="flex flex-col gap-[10px]">
+                <div className="w-[52px] shrink-0 text-body2-medium text-gray-08">폴더</div>
+                <FolderDrawer
+                  categories={categories}
+                  step={step}
+                  setStep={setStep}
+                  folder={folder}
+                  doc={doc}
+                />
+              </div>
+
+              <div className="flex flex-col gap-[10px]">
+                <div className="w-[52px] shrink-0 text-body2-medium text-gray-08">노트</div>
+                <DocumentDrawer
+                  categories={categories}
+                  step={step}
+                  setStep={setStep}
+                  folder={folder}
+                  doc={doc}
+                />
+              </div>
+            </div>
+
+            <div className="mt-[12px] line-clamp-2 text-body2-regular text-orange-06">
+              선택된 노트:{' '}
+              <span className="text-body2-bold">
+                {doc.list
+                  .filter((document) => doc.getCheckedIds().includes(document.id))
+                  .map((document) => document.name)
+                  .join(', ')}
+              </span>
+            </div>
+
+            <div className="mt-[31px] flex flex-col gap-[10px]">
+              <div className="w-[52px] text-body2-medium text-gray-08">퀴즈 수</div>
+              <QuizCountSelector
+                quizCount={quizCount}
+                setQuizCount={setQuizCount}
+                isPendingQuizCount={isPending}
+                maxQuizCount={maxQuizCount}
+              />
+            </div>
+          </div>
+
+          <div className="mt-[68px] flex flex-col items-center gap-[8px]">
+            <div className="text-center text-small1-regular">
+              <span className="text-gray-06">나의 별: </span>
+              <span className="text-gray-08">{userPoints}개</span>
+            </div>
+            <Button
+              variant="gradation"
+              className="flex w-[335px] gap-[10px] text-white"
+              onClick={() => {
+                handleCreateQuizzes({
+                  documentIds: doc.getCheckedIds() as number[],
+                  count: quizCount,
+                })
+              }}
+            >
+              <div>퀴즈 시작</div>
+              <div className="flex items-start gap-[8px] rounded-[16px] px-[10px] py-[3px]">
+                <Image src={icons.star} width={16} height={16} alt="" className="mt-px" />
+                <div className="text-text-bold">{quizCount}</div>
+              </div>
+            </Button>
           </div>
         </div>
-
-        <div className="mt-[12px] line-clamp-2 text-body2-regular text-orange-06">
-          선택된 노트:{' '}
-          <span className="text-body2-bold">
-            {doc.list
-              .filter((document) => doc.getCheckedIds().includes(document.id))
-              .map((document) => document.name)
-              .join(', ')}
-          </span>
-        </div>
-
-        <div className="mt-[31px] flex flex-col gap-[10px]">
-          <div className="w-[52px] text-body2-medium text-gray-08">퀴즈 수</div>
-          <QuizCountSelector
-            quizCount={quizCount}
-            setQuizCount={setQuizCount}
-            isPendingQuizCount={isPending}
-            maxQuizCount={maxQuizCount}
-          />
-        </div>
-      </div>
-
-      <div className="mt-[68px] flex flex-col items-center gap-[8px]">
-        <div className="text-center text-small1-regular">
-          <span className="text-gray-06">나의 별: </span>
-          <span className="text-gray-08">{userPoints}개</span>
-        </div>
-        <Button
-          variant="gradation"
-          className="flex w-[335px] gap-[10px] text-white"
-          onClick={() => {
-            handleCreateQuizzes({
-              documentIds: doc.getCheckedIds() as number[],
-              count: quizCount,
-            })
-          }}
-        >
-          <div>퀴즈 시작</div>
-          <div className="flex items-start gap-[8px] rounded-[16px] px-[10px] py-[3px]">
-            <Image src={icons.star} width={16} height={16} alt="" className="mt-px" />
-            <div className="text-text-bold">{quizCount}</div>
-          </div>
-        </Button>
-      </div>
-    </div>
+      )}
+    </Div100vh>
   )
 }
